feat: add table casting
Use a table type to resolve a foreign key into a nested object. Also works with an array of foreign keys.

This is done by calling the table type as a function, passing in a foreign key or an array of them, and doing so within a selector callback.

See the "with table cast" tests in `spec/sql.spec.ts` for examples